+++ Src/Messaging.Remote/Common/EmptyObject.cs
@@ -1,3 +1,6 @@
 ﻿namespace FastEndpoints;
 
-public sealed class EmptyObject { }
+public sealed class EmptyObject
+{
+    internal static EmptyObject Instance { get; } = new EmptyObject();
+}